# api/openapi.yaml
openapi: 3.0.3
info:
  title: Self Systems API
  version: 0.1.0
  description: Phase 1 backend API (skim-tier processing, unified chat commands, URL-only resources).
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/sync/events:
    post:
      summary: Publish sync event (bootstrap)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishSyncEventRequest'
      responses:
        "202":
          description: Event accepted for broadcast
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "400":
          $ref: '#/components/responses/BadRequestError'
  /api/v1/sync/offline-queue/enqueue:
    post:
      summary: Enqueue offline mutation for replay
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnqueueReplayMutationRequest'
      responses:
        "202":
          description: Offline mutation queued
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "400":
          $ref: '#/components/responses/BadRequestError'
  /api/v1/sync/offline-queue/replay:
    post:
      summary: Replay queued offline mutations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplayQueueRequest'
      responses:
        "200":
          description: Replay completed with summary
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "400":
          $ref: '#/components/responses/BadRequestError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/sync/conflicts:
    get:
      summary: List recorded sync conflicts
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: entity_id
          schema: { type: string }
          description: Optional entity filter.
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, default: 50 }
      responses:
        "200":
          description: Conflict records
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequestError:
      description: Request failed validation or payload parsing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnauthorizedError:
      description: Authentication is required or token validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Stable machine-readable error code.
          example: validation_error
...
    PublishSyncEventRequest:
      type: object
      required: [type, payload]
      properties:
        type:
          type: string
          enum:
            - sync.update
            - sync.reconnected
            - sync.resource.created
            - sync.resource.updated
            - sync.resource.deleted
            - sync.category.updated
            - sync.todo.updated
            - sync.reminder.updated
          example: sync.resource.updated
        payload:
          type: object
          additionalProperties: true
          description: Include entity_id for entity-scoped event types.
          example:
            entity_id: res-1
            changed_fields: [title, summary]
    EnqueueReplayMutationRequest:
      type: object
      required: [type, payload]
      properties:
        operation_id:
          type: string
          description: Optional client-generated idempotency key.
          example: op-123
        type:
          type: string
          enum:
            - sync.update
            - sync.reconnected
            - sync.resource.created
            - sync.resource.updated
            - sync.resource.deleted
            - sync.category.updated
            - sync.todo.updated
            - sync.reminder.updated
          example: sync.resource.updated
        occurred_at:
          type: string
          format: date-time
          description: Optional mutation timestamp. Defaults to server now.
        payload:
          type: object
          additionalProperties: true
          description: Must include entity_id for deterministic replay.
          example:
            entity_id: res-1
            title: Updated after reconnect
    ReplayQueueRequest:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          default: 100
...